requirements.txt
asgiref==3.3.1
Django==3.1.8
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
pytz==2021.1
sqlparse==0.4.1
black==21.6b0
pre-commit==2.13.0
django-rest-swagger==2.2.0
aenum==3.1.0
coverage==5.5
